fix: resolve ESM/CommonJS module compatibility issues in Next.js config
