+++ middleware.ts
@@ -11,7 +11,7 @@ export const config = {
   // Skip internal paths, API routes, and static assets
   matcher: [
     '/', // Match the root path
-    '/(en|es)/:path*', // Match locale-prefixed paths
+    '/(en|es|vi|pt)/:path*', // Match locale-prefixed paths (added pt for Portuguese)
     // Avoid matching internal Next.js paths, API routes, and static files
     '/((?!api|_next/static|_next/image|assets|favicon.ico|sw.js).*)'
   ]
