+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "NODE_OPTIONS='--no-deprecation' next build",
+    "build": "NODE_OPTIONS=--no-deprecation next build",
     "start": "next start",
     "lint": "next lint"
   },
